Add files via upload
Q1. What is Statistics?

Q2. Define the different types of statistics and give an example of when each type might be used.

Q3. What are the different types of data and how do they differ from each other? Provide an example of
each type of data.
Q5. Explain the concept of levels of measurement and give an example of a variable for each level.

Q6. Why is it important to understand the level of measurement when analyzing data? Provide an
example to illustrate your answer.

Q7. How nominal data type is different from ordinal data type.

Q8. Which type of plot can be used to display data in terms of range?

Q9. Describe the difference between descriptive and inferential statistics. Give an example of each
type of statistics and explain how they are used.

Q10. What are some common measures of central tendency and variability used in statistics? Explain
how each measure can be used to describe a dataset.