Add files via upload
Q1. What are the three measures of central tendency?

Q2. What is the difference between the mean, median, and mode? How are they used to measure the
central tendency of a dataset?

Q3. Measure the three measures of central tendency for the given height data:

[178,177,176,177,178.2,178,175,179,180,175,178.9,176.2,177,172.5,178,176.5]

Q4. Find the standard deviation for the given data:

[178,177,176,177,178.2,178,175,179,180,175,178.9,176.2,177,172.5,178,176.5]

Q5. How are measures of dispersion such as range, variance, and standard deviation used to describe
the spread of a dataset? Provide an example.

Q6. What is a Venn diagram?

Q7. For the two given sets A = (2,3,4,5,6,7) & B = (0,2,6,8,10). Find:

(i) A B

(ii) A B

Q8. What do you understand about skewness in data?
Q9. If a data is right skewed then what will be the position of median with respect to mean?

Q10. Explain the difference between covariance and correlation. How are these measures used in
statistical analysis?

Q11. What is the formula for calculating the sample mean? Provide an example calculation for a
dataset.

Q12. For a normal distribution data what is the relationship between its measure of central tendency?

Q13. How is covariance different from correlation?

Q14. How do outliers affect measures of central tendency and dispersion? Provide an example.