Add files via upload
Q1. What is the Probability density function?

Q2. What are the types of Probability distribution?

Q3. Write a Python function to calculate the probability density function of a normal distribution with
given mean and standard deviation at a given point.

Q4. What are the properties of Binomial distribution? Give two examples of events where binomial
distribution can be applied.

Q5. Generate a random sample of size 1000 from a binomial distribution with probability of success 0.4
and plot a histogram of the results using matplotlib.

Q6. Write a Python function to calculate the cumulative distribution function of a Poisson distribution
with given mean at a given point.

Q7. How Binomial distribution different from Poisson distribution?

Q8. Generate a random sample of size 1000 from a Poisson distribution with mean 5 and calculate the
sample mean and variance.

Q9. How mean and variance are related in Binomial distribution and Poisson distribution?

Q10. In normal distribution with respect to mean position, where does the least frequent data appear?