Add files via upload
Q1: What is Estimation Statistics? Explain point estimate and interval estimate.

Q2. Write a Python function to estimate the population mean using a sample mean and standard
deviation.

Q3: What is Hypothesis testing? Why is it used? State the importance of Hypothesis testing.

Q4. Create a hypothesis that states whether the average weight of male college students is greater than
the average weight of female college students.

Q5. Write a Python script to conduct a hypothesis test on the difference between two population means,
given a sample from each population.

Q6: What is a null and alternative hypothesis? Give some examples.

Q7: Write down the steps involved in hypothesis testing.

Q8. Define p-value and explain its significance in hypothesis testing.

Q9. Generate a Student's t-distribution plot using Python's matplotlib library, with the degrees of freedom
parameter set to 10.

Q10. Write a Python program to calculate the two-sample t-test for independent samples, given two
random samples of equal size and a null hypothesis that the population means are equal.

Q11: What is Student’s t distribution? When to use the t-Distribution.
Q12: What is t-statistic? State the formula for t-statistic.

Q13. A coffee shop owner wants to estimate the average daily revenue for their shop. They take a random
sample of 50 days and find the sample mean revenue to be $500 with a standard deviation of $50.
Estimate the population mean revenue with a 95% confidence interval.

Q14. A researcher hypothesizes that a new drug will decrease blood pressure by 10 mmHg. They conduct a
clinical trial with 100 patients and find that the sample mean decrease in blood pressure is 8 mmHg with a
standard deviation of 3 mmHg. Test the hypothesis with a significance level of 0.05.

Q15. An electronics company produces a certain type of product with a mean weight of 5 pounds and a
standard deviation of 0.5 pounds. A random sample of 25 products is taken, and the sample mean weight
is found to be 4.8 pounds. Test the hypothesis that the true mean weight of the products is less than 5
pounds with a significance level of 0.01.

Q16. Two groups of students are given different study materials to prepare for a test. The first group (n1 =
30) has a mean score of 80 with a standard deviation of 10, and the second group (n2 = 40) has a mean
score of 75 with a standard deviation of 8. Test the hypothesis that the population means for the two
groups are equal with a significance level of 0.01.

Q17. A marketing company wants to estimate the average number of ads watched by viewers during a TV
program. They take a random sample of 50 viewers and find that the sample mean is 4 with a standard
deviation of 1.5. Estimate the population mean with a 99% confidence interval.